IMAGE zoom on hover
What i did:

I created a new test file to try out the zoom on hover effect using overflow:hidden; and it worked perfectly first try.

This means that the problem is not overflow:hidden; but there is something else in the code affecting it. Furthermore, in chrome inspect element, the overflow:hidden; appears to be working so I am not sure why the effect doesn't work